Move files to annotation package
